chore: rename enum module.
