+++ conftool/model/enum.py
