+++ conftool/model/generation_interval.py
@@ -1,4 +1,4 @@
-from .config_enum import ConfigEnum
+from .enum import ConfigEnum
 
 class GenerationIntervalModel(ConfigEnum):
     HOURLY = 'Hourly'
