+++ conftool/model/grant_type.py
@@ -1,4 +1,4 @@
-from .config_enum import ConfigEnum
+from .enum import ConfigEnum
 
 class GrantTypeModel(ConfigEnum):
     PASSWORD = "password"
